experiments/regression/requirements_Sept2023.txt
absl-py==1.4.0
aim==3.17.5
aim-ui==3.17.5
aimrecords==0.0.7
aimrocks==0.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
array-record==0.4.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
base58==2.0.1
beartype==0.13.1
black==23.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
check-shapes==1.1.1
chex==0.1.82
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
datasets==2.13.1
decorator==5.1.1
deprecation==2.1.0
dill==0.3.6
distrax==0.1.4
dm-haiku==0.0.9
dm-tree==0.1.8
dropstackframe==0.1.0
einops==0.6.1
equinox==0.9.2
etils==1.4.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.99.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.5.4
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
gpjax==0.5.9
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
ipython==8.14.0
isort==5.12.0
jax==0.4.14
jaxkern==0.0.5
jaxlib==0.4.13
jaxlinop==0.0.3
jaxtyping==0.2.21
jaxutils==0.0.8
jedi==0.18.2
Jinja2==3.1.2
jmp==0.0.4
keras==2.12.0
kiwisolver==1.4.4
lark==1.1.7
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-collections==0.1.0
ml-dtypes==0.2.0
ml-tools @ git+https://github.com/vdutor/ml-tools@30f7983b93e77d23ced066cd2bedc537db551c4f
monotonic==1.6
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.7
-e git+ssh://git@github.com/vdutor/neural-diffusion-processes.git@df5dfd0881653fb9395d2c7daef679a9130203da#egg=neural_diffusion_processes
numpy==1.25.2
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.2.7
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.3.0
plum-dispatch==2.2.1
promise==2.3
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py3nvml==0.2.7
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
pyparsing==3.1.0
pytest==7.4.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
RestrictedPython==6.0
rsa==4.9
ruff==0.0.247
scipy==1.11.2
segment-analytics-python==2.2.3
simple-pytree==0.1.7
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
stack-data==0.6.2
starlette==0.27.0
tabulate==0.9.0
taskipy==1.12.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
tensorflow-cpu==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
tensorflow-probability==0.19.0
tensorstore==0.1.41
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.66.1
traitlets==5.9.0
typeguard==4.1.3
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.14.1
xmltodict==0.13.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.2
